Fix merge conflict in coreutils/timeout.c.

+++ coreutils/timeout.c
@@ -39,10 +39,11 @@
 //kbuild:lib-$(CONFIG_TIMEOUT) += timeout.o
 
 //usage:#define timeout_trivial_usage
-//usage:       "[-s SIG] SECS PROG ARGS"
+//usage:       "[-s SIG] [-k KILL_SECS] SECS PROG ARGS"
 //usage:#define timeout_full_usage "\n\n"
 //usage:       "Run PROG. Send SIG to it if it is not gone in SECS seconds.\n"
-//usage:       "Default SIG: TERM."
+//usage:       "Default SIG: TERM.\n"
+//usage:       "If it still exists in KILL_SECS seconds, send KILL.\n"
 
 #include "libbb.h"
 
@@ -54,6 +55,38 @@ static void kill_child(void)
 	if (child != INVALID_HANDLE_VALUE) {
 		kill_SIGTERM_by_handle(child);
 	}
+}
+
+/* Return TRUE if child exits before timeout expires */
+static NOINLINE int timeout_wait(int timeout, HANDLE proc, DWORD *status)
+{
+	/* Just sleep(HUGE_NUM); kill(parent) may kill wrong process! */
+	while (1) {
+		sleep1();
+		if (--timeout <= 0)
+			break;
+		if (WaitForSingleObject(proc, 0) == WAIT_OBJECT_0) {
+			/* process is gone */
+			GetExitCodeProcess(proc, status);
+			return TRUE;
+		}
+	}
+	return FALSE;
+}
+#else
+static NOINLINE int timeout_wait(int timeout, pid_t pid)
+{
+	/* Just sleep(HUGE_NUM); kill(parent) may kill wrong process! */
+	while (1) {
+		sleep1();
+		if (--timeout <= 0)
+			break;
+		if (kill(pid, 0)) {
+			/* process is gone */
+			return EXIT_SUCCESS;
+		}
+	}
+	return EXIT_FAILURE;
 }
 #endif
 
@@ -69,11 +102,13 @@ int timeout_main(int argc UNUSED_PARAM, char **argv)
 #endif
 	int parent = 0;
 	int timeout;
+	int kill_timeout;
 	pid_t pid;
 #if !BB_MMU
 	char *sv1, *sv2;
 #endif
 	const char *opt_s = "TERM";
+	char *opt_k = NULL;
 
 #if ENABLE_PLATFORM_MINGW32
 	xfunc_error_retval = 125;
@@ -83,7 +118,7 @@ int timeout_main(int argc UNUSED_PARAM, char **argv)
 
 	/* -t SECONDS; -p PARENT_PID */
 	/* '+': stop at first non-option */
-	getopt32(argv, "+s:" USE_FOR_NOMMU("p:+"), &opt_s, &parent);
+	getopt32(argv, "+s:k:" USE_FOR_NOMMU("p:+"), &opt_s, &opt_k, &parent);
 	/*argv += optind; - no, wait for bb_daemonize_or_rexec! */
 
 	signo = get_signum(opt_s);
@@ -93,6 +128,10 @@ int timeout_main(int argc UNUSED_PARAM, char **argv)
 	if (signo != SIGTERM && signo != SIGKILL && signo != 0)
 #endif
 		bb_error_msg_and_die("unknown signal '%s'", opt_s);
+
+	kill_timeout = 0;
+	if (opt_k)
+		kill_timeout = parse_duration_str(opt_k);
 
 	if (!argv[optind])
 		bb_show_usage();
@@ -128,17 +167,16 @@ int timeout_main(int argc UNUSED_PARAM, char **argv)
 		bb_daemonize_or_rexec(0, argv);
 		/* Here we are grandchild. Sleep, then kill grandparent */
  grandchild:
-		/* Just sleep(HUGE_NUM); kill(parent) may kill wrong process! */
-		while (1) {
-			sleep1();
-			if (--timeout <= 0)
-				break;
-			if (kill(parent, 0)) {
-				/* process is gone */
-				return EXIT_SUCCESS;
-			}
-		}
+		if (timeout_wait(timeout, parent) == EXIT_SUCCESS)
+			return EXIT_SUCCESS;
 		kill(parent, signo);
+
+		if (kill_timeout > 0) {
+			if (timeout_wait(kill_timeout, parent) == EXIT_SUCCESS)
+				return EXIT_SUCCESS;
+			kill(parent, SIGKILL);
+		}
+
 		return EXIT_SUCCESS;
 	}
 
@@ -163,21 +201,19 @@ int timeout_main(int argc UNUSED_PARAM, char **argv)
 
 	child = (HANDLE)ret;
 	atexit(kill_child);
-	while (1) {
-		sleep(1);
-		if (signo && --timeout <= 0) {
-			status = signo == SIGKILL ? 137 : 124;
-			break;
-		}
-		if (WaitForSingleObject(child, 0) == WAIT_OBJECT_0) {
-			/* process is gone */
-			GetExitCodeProcess(child, &status);
-			goto finish;
-		}
-	}
+	if (timeout_wait(timeout, child, &status))
+		goto finish;
+	status = signo == SIGKILL ? 137 : 124;
 
 	pid = (pid_t)GetProcessId(child);
 	kill(pid, signo);
+
+	if (kill_timeout > 0) {
+		if (timeout_wait(kill_timeout, child, &status))
+			goto finish;
+		kill(parent, SIGKILL);
+		status = 137;
+	}
  finish:
 	CloseHandle(child);
 	child = INVALID_HANDLE_VALUE;
